+++ react/EXPERIMENTAL_useTableBulkActions.ts
@@ -1,0 +1,3 @@
+import useTableBulkActions from './components/EXPERIMENTAL_Table/hooks/useTableBulkActions'
+
+export default useTableBulkActions