+++ react/components/EXPERIMENTAL_Table/Toolbar/InputSearch.tsx
@@ -1,7 +1,8 @@
 import React, { FC, FormEvent, InputHTMLAttributes } from 'react'
+import csx from 'classnames'
 
 import Input from '../../InputSearch/index'
-import { NAMESPACES } from '../constants'
+import { NAMESPACES, ORDER_CLASSNAMES } from '../constants'
 
 export type InputSearchProps = InputHTMLAttributes<HTMLInputElement> & {
   onSubmit: (event: FormEvent<HTMLFormElement>) => void
@@ -11,7 +12,7 @@ const InputSearch: FC<InputSearchProps> = ({ onSubmit, ...inputProps }) => {
   return (
     <form
       id={NAMESPACES.TOOLBAR.INPUT_SEARCH}
-      className="order-0 w-40"
+      className={csx(ORDER_CLASSNAMES.TOOLBAR_CHILD.INPUT, 'w-40')}
       onSubmit={onSubmit}>
       <Input {...inputProps} />
     </form>